refactor: rewrite AddIcon to AddIconByType
Huge rewrite of .htaccess file so that AddIcon are rewritten to AddIconByType for most of the file. Icon are now displayed on a real media type basis <hich is more reliable that file extension.

More extensions are now supported.

File extensions that are not associated to a media type are still using the old AddIcon syntax.